edited authentication where we log out and it returns the user back to the home page
