+++ client/src/App.js
@@ -2,11 +2,6 @@ import React from "react";
 import { useGlobalContext } from "./context/GlobalContext";
 import AuthenticatedApp from "./components/AuthenticatedApp";
 import UnauthenticatedApp from "./components/UnauthenticatedApp";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import Welcome from "./pages/Welcome";
-import Profile from "./pages/Profile";
-import History from "./pages/History";
-import Login from "./pages/Login";
 
 function App() {
   const [state, dispatch] = useGlobalContext();
