+++ client/src/pages/History.js
@@ -1,1 +1,7 @@
 //user can check the full list of posts
+import React from "react";
+
+const History = () => {
+  return <div>This is our history page</div>;
+};
+export default History;
